requirements.txt
Flask==3.1.0
Pillow==11.1.0
Requests==2.32.3
torch==2.1.0+cpu
torchvision==0.15.2+cpu
transformers==4.47.1
TTS==0.22.0